Create endpoint for tags
